Move html_entities and html_parser moves for genshi._six to genshi.compat.
